requirements.txt
asgiref==3.5.2
astroid==2.8.5
asttokens==2.2.1
backcall==0.2.0
black==22.12.0
certifi==2022.12.7
click==8.1.3
colorama==0.4.4
decorator==5.1.1
distlib==0.3.6
Django==4.1.4
django-cors-headers==3.13.0
django-livereload-server==0.4
django-livesync==0.5
djangorestframework==3.14.0
executing==1.2.0
filelock==3.8.2
fire==0.1.2
ipython==8.7.0
isort==5.10.1
jedi==0.18.2
lazy-object-proxy==1.6.0
matplotlib-inline==0.1.6
mccabe==0.6.1
mypy-extensions==0.4.3
optional-django==0.1.0
parso==0.8.3
pathspec==0.10.3
peewee==2.8.8
pickleshare==0.7.5
pipenv==2022.11.30
platformdirs==2.4.0
prompt-toolkit==3.0.36
pure-eval==0.2.2
Pygments==2.13.0
pylint==2.11.1
pytube==12.1.0
pytz==2022.7
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.6.2
toml==0.10.2
tomli==2.0.1
tornado==6.2
traitlets==5.6.0
types-beautifulsoup4==4.11.6.1
typing_extensions==4.0.0
tzdata==2022.7
virtualenv==20.17.1
virtualenv-clone==0.5.7
virtualenvwrapper-win==1.2.7
watchdog==2.2.1
wcwidth==0.2.5
websocket-client==1.4.2
Workon==0.8.2
wrapt==1.13.3
